+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\index.js":"1","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\reportWebVitals.js":"2","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\App.js":"3","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\aboutUs\\AboutUs.js":"4","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\home\\Home.js":"5","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\allPlants\\AllPlants.js":"6","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\components\\header\\Header.js":"7","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\context\\LanguageContext.js":"8"},{"size":517,"mtime":1617897365773,"results":"9","hashOfConfig":"10"},{"size":375,"mtime":1617897365803,"results":"11","hashOfConfig":"10"},{"size":4009,"mtime":1617906639247,"results":"12","hashOfConfig":"10"},{"size":594,"mtime":1617906603028,"results":"13","hashOfConfig":"10"},{"size":485,"mtime":1617906469351,"results":"14","hashOfConfig":"10"},{"size":558,"mtime":1617906960866,"results":"15","hashOfConfig":"10"},{"size":1483,"mtime":1617907832298,"results":"16","hashOfConfig":"10"},{"size":91,"mtime":1617904036726,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"13frhnf",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"34"},"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\index.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\reportWebVitals.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\App.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\aboutUs\\AboutUs.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\home\\Home.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\allPlants\\AllPlants.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\components\\header\\Header.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\context\\LanguageContext.js",[],["35","36"],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\index.js":"1","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\reportWebVitals.js":"2","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\App.js":"3","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\aboutUs\\AboutUs.js":"4","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\home\\Home.js":"5","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\allPlants\\AllPlants.js":"6","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\components\\header\\Header.js":"7","D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\context\\LanguageContext.js":"8"},{"size":654,"mtime":1617964414604,"results":"9","hashOfConfig":"10"},{"size":375,"mtime":1617897365803,"results":"11","hashOfConfig":"10"},{"size":3510,"mtime":1617965353368,"results":"12","hashOfConfig":"10"},{"size":594,"mtime":1617906603028,"results":"13","hashOfConfig":"10"},{"size":446,"mtime":1617967950595,"results":"14","hashOfConfig":"10"},{"size":558,"mtime":1617906960866,"results":"15","hashOfConfig":"10"},{"size":1799,"mtime":1617967898657,"results":"16","hashOfConfig":"10"},{"size":681,"mtime":1617964549379,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"13frhnf",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\index.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\reportWebVitals.js",[],["35","36"],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\App.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\aboutUs\\AboutUs.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\home\\Home.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\pages\\allPlants\\AllPlants.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\components\\header\\Header.js",[],"D:\\front-end lessen\\React\\Les 19 - 8 april\\Context\\src\\context\\LanguageContext.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
