--- src/components/FlagButton.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function FlagButton({ FlagNL }) {
-    return (
-        <button type="button">
-
-        </button>
-    )
-}
-
-export default FlagButton;
