+++ src/components/header/Header.css
@@ -54,6 +54,14 @@ nav a:link {
     margin-left: 10px;
 }
 
+button {
+    width: 30px;
+    border: none;
+    text-decoration: none;
+    outline: none;
+    background-color: transparent;
+    cursor: pointer;
+}
 
 .active {
     color: red;
